+++ CHANGELOG.rst
@@ -6,13 +6,18 @@ Change log
 `Next version`_
 ~~~~~~~~~~~~~~~
 
+
+`0.7`_ (2018-10-18)
+~~~~~~~~~~~~~~~~~~~
+
 - Made error reporting in ``process_imagefields`` include more info.
 - Made image field validation catch errors while determining the image
   dimension too.
 - Fixed a problem where older versions of Django didn't allow specifying
   the chunk size for iterating over querysets.
 - Modified django-imagefield's internals to allow changing the type and
-  extension of generated images.
+  extension of generated images by way of dynamically specifying the
+  processing pipeline.
 - Changed the API of the ``get_image`` callable in processors to only
   return the image without the context (since the context is mutable and
   available already).
@@ -110,4 +115,5 @@ Change log
 .. _0.4: https://github.com/matthiask/django-imagefield/compare/0.3...0.4
 .. _0.5: https://github.com/matthiask/django-imagefield/compare/0.4...0.5
 .. _0.6: https://github.com/matthiask/django-imagefield/compare/0.5...0.6
-.. _Next version: https://github.com/matthiask/django-imagefield/compare/0.6...master
+.. _0.7: https://github.com/matthiask/django-imagefield/compare/0.6...0.7
+.. _Next version: https://github.com/matthiask/django-imagefield/compare/0.7...master